Integer <=> Float

+++ db/goals_db.py
@@ -1,5 +1,5 @@
 from sqlalchemy import Column, ForeignKey
-from sqlalchemy import Integer, String, Date
+from sqlalchemy import Integer, String, Date, Float
 import datetime
 from db.db_connection import Base, engine
 
@@ -7,8 +7,8 @@ class GoalsInDb(Base):
     __tablename__ = "goals"
     name        = Column(String, primary_key=True)
     username    = Column(String, ForeignKey("users.username"))
-    current_val = Column(Integer)
-    final_value = Column(Integer)
+    current_val = Column(Float)
+    final_value = Column(Float)
     final_date  = Column(Date)
 
 Base.metadata.create_all(bind=engine)
